requirements.txt
fastapi[all]==0.95.0
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy==2.0.9
psycopg2==2.9.6
pydantic==1.10.7